+++ libraries/teardown/CHANGELOG.md
@@ -7,6 +7,10 @@ The change log is available [on GitHub][2].
 [1]: http://semver.org/spec/v2.0.0.html
 [2]: https://github.com/roman/Haskell-teardown/libraries/teardown/CHANGELOG.md
 
+## Unreleased
+
+* Add NFData instance for `TeardownResult` record
+
 
 ## v0.1.0.1